+++ docs/examples/userschema/get.ts
@@ -1,0 +1,8 @@
+import { Client } from '@robgietema/nick';
+
+const cli = Client.initialize({ apiPath: 'http://localhost:8080' });
+const login = await cli.login({ username: 'admin', password: 'admin' });
+
+const { data } = await cli.getUserSchema({
+  token: login.data.token,
+});